ARM: sun6i: select mmc order based on VER_REG
selecting the mmc boot device based on mmc content can fail if the
device is booted from eMMC and there is a SD-card with u-boot plugged
in.
On sun6i we can ѕelect the correct boot device based on the hardware
configuration pins in BOOT_SEL. We only want to do this if UBOOT_SEL is
not active since this overrides the BOOT_SEL settings to SD card.

Signed-off-by: Klaus Goger <klaus.goger@theobroma-systems.com>
Signed-off-by: Jakob Unterwurzacher <jakob.unterwurzacher@theobroma-systems.com>