drivers: USB: OHCI: make 64-bit safe
OHCI has a known limitation of allowing only 32-bit DMA buffer
addresses, so we have a lot of u32 variables around, which are assigned
to pointers and vice versa. This obviously creates issues with 64-bit
systems, so the compiler complains here and there.
Many U-Boot targets (including sunxi) are limited to 32-bit memory
anyways, so to avoid more invasive code changes we just adjust some
casts and types and assume that the EDs and TDs are in the lower 4GB.
This fixes compilation of the OHCI driver for the Pine64.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>