sunxi: Scan DT tree node '/clocks' on sunxi boards
Moving to a DT-based clock framework on sunxi requires the subnodes
of /clocks to be bound on boot-up.  As /clocks does not contain a
compatible-string, the U-Boot DT parsing code ignores it (and any
subnodes).  To overcome this limitation, the sunxi board-init code
retrieves the /clocks node and explicitly starts a DM scan of the
subnodes.

Signed-off-by: Philipp Tomsich <philipp.tomsich@theobroma-systems.com>