cmd: sspi: Add an additional argument to set the bus frequency
The spi command is rather useful for board verification, bring-up and
software testing.  One use in our lab is the testing of the clock
code, which requires the ability to send data at different SPI clocks.

The changeset adds an additional (optional) positional argument to the
'spi' command that allows the selection of a SPI clock for
communication (the previous default of 1MHz is retained, if the
argument is omitted).

Signed-off-by: Philipp Tomsich <philipp.tomsich@theobroma-systems.com>