sun8i_emac: update to work with pinctrl-sunxi, reset-sunxi and clk-sunxi
parse_phy_pins() is parsing the pinctrl-entry of the emac-node, which
is in conflict with the new DM pinctrl driver for sun50i.  So when
building for sun50i and if the pinctrl driver is active, let the
parse_phy_pins()-function do nothing.

Use the DM reset and DM clk frameworks for getting the emac module out
of reset and enabling the clock gates.

Signed-off-by: Philipp Tomsich <philipp.tomsich@theobroma-systems.com>